Task-2/requirements.txt
tensorflow
streamlit
numpy
pandas
matplotlib
pillow